Rename folder, minor import fix
