+++ opendm/video/checkers.py
